pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["Nazmus Saadat <nazmus.saadat1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.3"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
pandas = "^2.2.2"
pydeck = "^0.9.1"
pygithub = "^2.3.0"
flask = "^3.0.3"
levenshtein = "^0.25.1"
reflex = "0.5.3"

[tool.poetry.scripts]
run = "hackathon.src.run:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
